refactor helper functions into utils
